+++ impl-java/src/babashka/impl/reify2/interfaces.clj
@@ -47,4 +47,5 @@
                  org.jline.reader.Highlighter
                  org.jline.reader.Parser
                  org.jline.reader.ParsedLine
-                 org.jline.reader.Widget])
+                 org.jline.reader.Widget
+                 org.jline.terminal.Terminal$SignalHandler])